+++ CHANGELOG.md
@@ -98,6 +98,7 @@ hdhomerun:
 - A new filter field `caption` has been added. This field is used to bypass the `title/name` issue. 
 If `caption` is provided, its value is read from `title` if available, otherwise from `name`. 
 When setting `caption`, both `title` and `name` are updated.”
+- Counter has now an attribute padding. Which fills the number like 001.
 
 # 2.2.5 (2025-03-27)
 - fixed web ui playlist regexp search